restructure files & added Router
