+++ src/components/common/Canvas/Canvas.scss
