+++ src/components/common/index.ts
@@ -1,4 +1,4 @@
-import App from './App/App';
+import App from '../App/App';
 import Canvas from './Canvas/Canvas';
 import TestChat from './TestChat/TestChat';
