+++ src/components/pages/HomePage/HomePage.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const HomePage = (): React.ReactElement => {
+    const a = '123';
+    return <h1>hello world</h1>;
+};
+
+export default HomePage;
