+++ src/components/pages/NotFoundPage/NotFoundPage.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const NotFoundPage = (): React.ReactElement => {
+    return <h1>404 NOT FOUND!!</h1>;
+};
+
+export default NotFoundPage;
