+++ src/components/pages/RoomPage/RoomPage.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const RoomPage = (): React.ReactElement => {
+    const a = '123';
+    return <h1>room world</h1>;
+};
+
+export default RoomPage;
