+++ src/components/pages/index.ts
@@ -1,0 +1,5 @@
+import HomePage from './HomePage/HomePage';
+import RoomPage from './RoomPage/RoomPage';
+import NotFoundPage from './NotFoundPage/NotFoundPage';
+
+export { HomePage, RoomPage, NotFoundPage };
