+++ src/hooks/useDispatch.ts
@@ -1,4 +1,4 @@
 import { useDispatch as ogUseDispatch } from 'react-redux';
-import { AppDispatch } from '../../store/store';
+import { AppDispatch } from '../store/store';
 
 export default () => ogUseDispatch<AppDispatch>();
