+++ src/interfaces/board.tsx
