+++ src/interfaces/webSocket.tsx
